Address installation confusions
The workflow of installing the plugin had a few confusions which would not have worked in the long term. Added the path to the actual repository so sdist  will not fail. Added install.bash to standardize the installation process without a lot of copying/pasting. Updated the README.MD with the new information.