Renamed plugin and updated install script, dependencies and README.md
Fixed vulnerability found in package.json for eslint and upgraded remaining dependencies. Changed name of the plugin from sample-plugin to policies-plugin. Changed the update.bash script and enabled files to use the new directory and package names. Updated the README.md to show the current package and plugin names and added a quick install section to the top.