Plugin now shows all effective policies as well as default rules
A dict was created to provide Description, Operations, Scopes and Default rules for policies. This is because the function used from oslo.policy that provides the list of all the effective policies, both default and file policies, doesn't provide those values.
Policy model class needs clean up after the new changes and the keystone dict file could be moved to its own folder. The roles and the differences between the api and the client should be discussed and agreed upon in order to distinguish which operations belong where.