Fixed problem with search not showing results from other pages
Removed Horizon's own implementations and used vanilla angularjs to achieve these because they didn't expose the necessary parameters to make the changes we wanted and thus conflicted with our implementation. We can take a closer look at Horizon's angularjs widget implementations at a later date but for now there is no need for it. This is because they don't have implementations for everything we need and mixing regular angularjs with their ready made widgets resulted in a lot of important functionality getting broken with no easy or clear way to fix them.
In this push, I had to implement custom functionality to create the column sorting along with a custom pagination implementation. The pagination could be visually improved but it works as it should. A new functionality has been added to set the number of table rows to show at a time which is useful for page load time.
The search functionality has been also recreated and a custom filter has been created so that the search is done on the whole data set and not just the ones displayed in the current page of the table.