Made structural changes to back end in preparation for policy retrieval
The api and the client have been switched as the api is the component which angularjs communicates with and the client is a client for keystone or more specifically oslo.policy. This change makes the naming of each of the two components better resembling of their functional purpose.
The back end now supports retrieval of single policy rather than just all the policies at once. However, the implementation in this commit uses a hard coded URL and policy name which needs to be changed so it can accept parameters instead. The reason for pushing this incomplete commit is so that the full team can work on the new changes for this sprint.
Further, the angularjs filter module has been removed as it is not deemed necessary. Rickard has pointed out an implementation that doesn't require the use of a custom filter. The change can be seen in the table.html file in the ng-repeater-start tag where the startFilter was replaced.