Single policy retrieval works in both back end and front end
The back end now accepts URL passed value for the target of the policy and returns the requested policy. In the table, clicking on a target result retrieves its policy from the back end and renders the values underneath the table. The next step is to move the values of the requested policy so that they are rendered in a popup window.