Implemented policy details pop up modal
Clicking on a policy's target in the policies table instantiates a modal overlay that displays the details from the selected policy. This will be used to make modifications to policies. The details.html file and the DetailsController can be moved to their own child directory and imported as a module.