Added front end functionality to modify policy values in details modal
The details modal can now make modifications to a selected policy. Extra functionality has been added to enable editing of policy fields as well as buttons to add and remove items in the policy's scopes and operations arrays.
The front end currently performs all modifications directly as they happen rather than after pressing the save button. This behavior needs to be changed before connecting it to the back end policy modification methods. It has been left in this commit for testing purposes.