Added so that backend sends back the modified policy when refreshing
The modified policy that is sent to backend are now overwriting the default rule with the modified policy, and the new rule is shown in the panel. Unfortunately this is done with a output-file at /etc/keystone/. I can't see any other way around it, yet. Also commented out all Edit/Remove buttons except 'Edit' in the Rule, since this is the only modification that the plugin can handle at the moment.