Added policy modifications notification toasts and changed API endpoint
The setPolicy method now calls the /policy url endpoint instead of the /policies url endpoint. The back end now returns a status code of 200 and it is used to display a toast success notification. Functionality needs to be added to test and return different status codes depending on operation result.