Moved details controller and modal html file to its own module
The modal instance which is opened when clicking on a policies target in the policies table has been moved to its own module along with the controller responsible for its data. This has been done to separate different functionalities into their own modules and to make them easier to work with.