Fixed problem with policies table search & removed unnecessary functions
The model used by the search bar to filter the policy rows in the policies table was accidentally omitted during one of the previous modifications to it. It has been added back to it and the search functionality works again now. Some commented out code in the policy modal has been removed along with the functions that were called within it. This is due to the fact that the policy.json file doesn't take any scope, operations or description values and thus modifications to those fields is not possible.