Added tool-tips and character limits based on number of visible columns
The number of characters shown in each cell when its row is not expanded is now set by a function that takes into account the number of columns that are visible in the table. This helps keep a uniform row height for collapsed rows in the policies table. Cells whose values exceed the character limit are visually indicated with three dots (...) where the text is cut off. Tool-tips have been added to quickly display the full cell's value when it exceeds the character limit or when not all lines are visible when its row is collapsed.