Fixed problem with column width not being set correctly when resized
The problem was due to the fact that the resize handler exited within an element with an ng-click directive, which lead to the parent element's ng-click getting fired when the column resize handle is clicked. The ng-click triggered a re-sorting of the row order on mouse click release which interferes in the column width setting operation.
To solve this, the ng-click has been set to the name of the column rather than the whole column, margins around the column header have also been added to prevent accidental triggering of the sort functionality when attempting to resize a column's width.