Changes to column widths & visible columns are now saved in localstorage
Two scope variables have been added, one for visible columns and one for column widths. The values for those are loaded from the browser's local storage on page load. Changes to visible columns are automatically stored using an ng-changed directive in the policy module's html code, more specifically in the table settings drop down menu check-boxes responsible for showing and hiding different columns. As for the column widths, several $watch scoped methods have been created for observing the width of each column. This might be a bit excessive and could probably be worked on to find a lighter implementation. As it stands, the newly Observable API hasn't been widely implemented across browsers and as such doesn't meet the browser compatibility requirements for the plugin by Horizon.
Some leftover functions and variables have also been removed in this commit.