Searching in specific columns works now
A new custom filter was created for this purpose. It checks which column search check boxes are selected and limits the search only to those columns.
The previous 'query' variable has been replaced with 'search'. The getData method used for counting the numbers of pages and the watcher method for resetting the current page to 0 when the search query changes have also been updated.