Fixed details and editor modals and now saving changes works in both
Recreated details modal so that it does not apply changes directly to the scoped variable but instead to a controller variable. When the save button is pressed, the value of the scoped variable is assigned the value of the changed controller variable.
Fixed problems with errors when pressing save in the editor modal and html elements calling the wrong model.