Added Editor toolbar buttons
The buttons have been added but they haven't been connected to their respective functions yet. The upload and download buttons can be removed from the bottom of the page while the save and cancel buttons can remain there. The chose file button for the upload functionality can be shown and hidden when needed using a transition animation that is triggered by pressing the button.