Restructured project structure and separated data model from controllers
A new services has also been created as single source of truth to handle the data model and share access to it across the different controllers. Performance in this implementation lives up to the expectations and it has decreased the number of used $scopes $watchers.

New components have been created for the different functionalities in the plugin. This has been done in order to better adhere to the mvc spirit and to make testing easier.

Some new functionalities have also been added such as the selected policies manager and the custom search field. Those functionalities appear with a transition animation according to specified conditions such as toggling a button or selecting policies.

Some code remains to be added such as storing and restoring to and from browser local storage for values such as columns visibility and width as well as set policy methods that connect the view to the Api.