Recreated quick actions as reusable actions & added context menu actions
* Changes in this commit:

- Moved policies-editor quick actions to their own modules under their own top level actions directory and recreated them as providers so that they can be reused across the applications. Examples of such actions are: upload, download, copy, print, etc...

- Moved Policies Service to top level module Models and renamed it to PoliciesModel. The policies object inside it has been renamed to data. The reasoning behind this move is to distinguish between it as a data model service for policies and the policies module which holds the main view.

- Added a context menu module which allows the user to use quick actions through right clicking a policy item in the policies table. Those context menu actions currently include: toggle select policy, toggle expand policy, quick edit policy, edit policy inn editor, copy policy, download policy, print policy and restore policy to default value.

* Implemented following functionalities: 
- Select all currently shown policies in the policies-table.
- Deselect all currently shown policies in the policies-table.
- Deselect all selected policies.
- Print selected policy.
- Edit selected policy in form editor.
- Edit selected policy in text editor.
- Copy selected policy to clipboard.

* UI improvements:
- Fixed the design and the transition of the custom search bar.
- Fixed the design and the transition of the table actions bar.

* Known bugs:
- The checkbox-model currently contains a bug that shows unchecked policies-table items as checked when they are not.
This can be recreated by checking a policy item and then navigating to a different page, the item in the same position across the different pages will also be checked when it actually isn't.
This seems like a UI bug and the view isn't properly being updated. It also seems that it has to do with the use of the 'track by $index' property of the ng-repeat of the policies table.
The checkbox-model should use the same property when track by is used but finding a way to implement an $index value in a policy hasn't been successful so far. Removing the 'track by $index' from ng-repeat seems to fix the problem but at the cost of performance.
When 'track by $index' is not specified the page load time when navigating between pages of the table goes up to about 0.2 to 0.3 seconds. This is because angularjs renders items not in the view as dirty and recreates them on each navigation which seems to fix the view bug but it comes at a huge cost. A better fix or checkbox model implementation needs to be implemented.

- The toggle expand functionality in the context menu seems to reset the current page to the first page but it is otherwise working. The behavior should be looked into as it shouldn't change the current page.