Further changes to test configuration files in order to start testing
The dependency issue seems to have been fixed. The only issue remaining seems to be that tox complains about ''../horizon' not being an install-able location and that setup.py is not found' which seem like a path configuration issue. The command that was used to run the test is:
tox -e karma-local