Fixed table item selection and expansion bugs and other table UI bugs
The sort functionality has been moved to the policies table for more control and better performance when performing actions in current view.
There is currently an issue with sorting by default rule, scopes and operations but it will be fixed in the next commit after changing the name of the default rule from 'default' to 'defaultRule' application wide in order to avoid reserved naming issues.
The checkbox-model has been removed and replaced with the necessary functionalities in the policies table controller.